Defined Data Classes
- introduced two new classes in gallery-core subdirectory, namely:
Album and Picture.
- The Album class will take care of added new Albums: their name and Id.
- The Picture class will take care of added new Pictures: their id,
album's id of which they are part of, their location on the disk.
- This system design method (of keeping albums and picture class) is
delibrately chosen so as to avoid confusion of wether to load an album
empty or with all its picture elements. In our case the album will be
loaded empty, and the pictures can be loaded as need permits.
-The getters and setters for both the classes' members have been defined
appropriately.